add scroll index, edit, delete, protect edit route, add scroll toolbar, add user page

+++ src/routes/ScrollDelete.jsx
@@ -1,9 +1,18 @@
 import { redirect } from 'react-router-dom';
-import { deleteScroll } from '../services/scrolls';
+import { deleteScroll, getScroll } from '../services/scrolls';
+import { getUser } from '../services/users';
 
 export async function action({ params }) {
-  console.log(`hitting scrollDeleteAction`);
+  console.log(`hitting scrolldelete action`)
+  const scroll = await getScroll(params.scrollId);
+  console.log(scroll.createdBy._id)
+  const user = await getUser();
+  console.log(user._id)
+  if (!user || user._id !== scroll.createdBy._id) {
+    throw redirect(`/scrolls/${params.scrollId}`);
+  }
+  console.log(`user is the author. preparing to delete`)
   await deleteScroll(params.scrollId);
-  console.log(`now it should redirect`)
+  console.log(`deleted. preparing to redirect`)
   return redirect('/');
 }
